[Refactor] #929 Moved mon_take_hit() from monster-status.cpp/h to monster-damage.cpp/h
